[CS] Less output at end of imports
Now that we have the results database and Metis starting to work
locally, getting all validation errors printed out (being unbounded in
size and without even being valid edn to copy into a text editor or repl
to investigate) has outlived its usefulness. Curbs the output to the
location of the temp sqlite db, the xml file, and the import id that can
be used to find validation errors in the results db.

Cuts the final ctx from the error, as the stacktrace and initial context
are what you really want.