[CS] Insert all the data into the database
When inserting multiple rows into a database with korma, the keys of the
first row determine which columns get values. So if a later row has a
key the first one doesn't, the associated value will not be inserted
into the database! This is very bad!

This is only an issue with XML importing, since during a CSV import
each row (by virtue of how they're constructed) will share all keys with
every other row.

This patch updates `bulk-import` to make all rows share all keys. Only
the first row could be updated, but I'd like to protect myself if korma
started using, say, the last row to figure out which columns to bother
with.