Add simple_type to xml_tree_values
simple_type is a version of the path for a value without the index
numbers, to aid in faster querying. Finding all StreetSegment ids (for
example) can be an equality check for a simple_path of
`VipObject.StreetSegment.id` instead of using an lquery with wildcards,
like `VipObject.0.StreetSegment.*{1}.id`. The deeper the nesting, the
greater the payoff.