Fix v5 stats query
Previously, it was querying for paths which *didn't* have particular
values in the third part of the path. However, feeds that included
extraneous elements or had attributes on the VipObject would result in
an attempt to insert values into non-existent columns, since the insert
query is built from the third path parts.

Now it specifically queries for only those elements covered by the
v5_statistics table.