Collect and process validation errors async
When a validation error is reported, instead of storing it on the
context, which balloons and balloons as time goes on, instead, put it on
a core.async channel. That channel is taken from and every so
often (once a number of messages is received or a timeout reached), the
validation errors are committed to the database.

Because some errors are gathered before the specification version is
known (and thus which table the errors must go to), the errors aren't
processed until the spec version is decided. This was accomplished by
turning the spec-version into an atom with a watch.

This is a big one, and a long time coming. This may result in a very
slight performance boost, but will definitely alleviate some OutOfMemory
issues with particularly bad feeds.