Use only 1 connection to the DB.
`lazy-select-xml-tree-values` uses a cursor, which in PostgreSQL, is
implemented per-processes; since each connection gets a new server
process, this means that if we use a connection pool we will also have
to land on the proper connection in order to find it. c3p0, the
connection pooling library that Korma uses, doesn't expire connections
by default [1], so we'll always be working with the connection on which
the cursor was created.

[1]: http://www.mchange.com/projects/c3p0/#managing_pool_size