Provide categories for crosstab; adjust street segment errors
When a category's value is null, crosstab shifts values to the
left (e.g., a null 'type' would put the 'id' in the 'type' slot);
providing a category query fixes nulls in the proper slot.

Counting the number of street segment errors was inefficient, but if we
reconsider the query, we can take advantage of the fact that we already
know the paths for any locality. To make that work, we need to reorder
the pipeline by building summaries earlier.

We also found that the ltree path matching operator will match
'VipObject.*{1}.StreetSegment.*{1}' about 10 times faster than
'*.StreetSegment.*' on pretty large datasets, and when we add a gist
index on the path, it cuts roughly in half again.