change column types in the sqlite street-segments table
So we often get strings in these fields, and it's
a "quirk" of sqlite that it doesn't care. Only,
it seems to be related to a seg fault happening
with the upgraded sqlite-jdbc dependency.

It seems ok to make this change since the postgres
version of the same table uses TEXT fields too. And
it made the WA feed process finally.