Biggish refactor
There are several interrelated refactors going on
in this commit:

* Was getting conflicts between java.io.File and
java.nio.file.Path in a number of locations (in
one pipeline run something would be a File, in
another it would be a Path), so I standardized
everything to be a Path when stored in the
context, dropping to File only when needed. Still
not entirely sure this was a better thing to do
overall, but at least now it's consistent.

* The `:input` keyword in the context was being
constantly swapped for different types of things
in the first part of the pipeline (EDN message,
file, extracted files, etc). This has been replaced
with specific keywords for each of the steps, and
the start of the pipeline has been altered a bit
so as to receive an initial context rather than
something merged into the initial context at
`:input`. This also allowed me to pull out some
processing flags that had gotten put in at the top
of the pipeline and now they get passed in the
initial-context. Future improvements planned will
be to add some specs for pre/post validations of
pipeline steps that expect a certain context shape
incoming and alter the shape outgoing.

That's the bulk of this commit. See the changes in
the README for more.