Form-encode the generated database connection string
While this technically is not URL encoding, this is also the method of
decoding that some (most?) JDBC drivers (checking PostgreSQL and MySQL,
but probably others) use, and so it will work for most use cases.
Critically, hand-written parsers seen in JDBC driver code typically look
for /, &, and =, so as long as these are encoded using a method that
java.net.URLDecoder can decode, everything will work nicely.