Move locality stats loop to Clojure code
This refactor allows us to see the progress of
the locality stats computations by moving the
loop outside of a database function and into
clojure code. It pays a slight performance penalty
to do so, but gives good insight into one of the
longest running and opaque steps of the 5.X
pipelines.

Possible future improvement could be to batch
process instead, so we send up to 10 locations to
be computed at a time, unless there are fewer than
10 locations or something.

Hopefully this information will do two things:
1) reassure us that progress is happening in feed
   processing
2) potentially hint at localities that take a long
   time and lead us to performance improvements